Bootstrap setting restrictions resolved
In readonly mode, the bootstrap value cannot be changed.
Therefore when a bootstrapswitch value needs to changed
but the switch itself needs to remain in readonly, the
settings go ignored.
This is causing a mismatch in displayed information and
the crimson filetree.

Resolved by putting the switch out of readonly mode when
a change of its state is made. Then set it back to its
old state.