Change dac dither mixer control label. Enabling of dither controls.
The dac_dither_mixer_enable label used to be called Dither Mixer. It's not entirely clear
from that label what the function is, so I changed it to "Dither Up-Conversion (+Fs/2)".

Also, previously, the dither controls would be modifiable even when dither was not enabled.
As requested, dither controls were disabled / read-only if dither was not enabled.